feature: add Tutorial Model
